chore(frontend): rename service directory to services
