+++ apps/frontend/services/auth.ts
