+++ apps/frontend/services/base.ts
